ref(manifests): Clearer manifest types and documentation
This moves the manifests to their own `mod` with additional documentation.
I also fixed some issues with places where we needed a map instead of a
concrete struct
