+++ oam/petclinic.yaml
@@ -91,7 +91,7 @@ spec:
           properties:
             target: httpserver
             values:
-              address: 0.0.0.0:8080
+              address: "0.0.0.0:8080"
 
     - name: httpserver
       type: capability
