+++ oam/simple1.json
@@ -44,7 +44,7 @@
                         "properties": {
                             "target": "webcap",
                             "values": {
-                                "port": 8080
+                                "port": "8080"
                             }
                         }
                     }
@@ -75,7 +75,7 @@
                                 {
                                     "name": "haslights",
                                     "requirements": {
-                                        "ledstatus": "enabled"
+                                        "ledenabled": "true"
                                     }
                                 }
                             ]
