+++ oam/simple1.yaml
@@ -28,7 +28,7 @@ spec:
           properties:
             target: webcap
             values:
-              port: 8080
+              port: "8080"
 
     - name: webcap
       type: capability
@@ -50,5 +50,5 @@ spec:
             spread:
               - name: haslights
                 requirements:
-                  ledstatus: enabled
+                  ledenabled: "true"
                 # default weight is 100
