+++ src/events/mod.rs
@@ -1,6 +1,5 @@
 mod data;
 mod deser;
-mod parser;
 mod ser;
 mod types;
