+++ src/lib.rs
@@ -3,6 +3,7 @@ use std::time::Duration;
 pub mod commands;
 pub mod consumers;
 pub mod events;
+pub mod model;
 pub mod nats_utils;
 pub mod storage;
 pub mod workers;
